+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Allow adding custom annotations to created App CRs.
+- Set "config-controller.giantswarm.io/version" label to "0.0.0" on created App
+  CRs.
+
 ## [3.5.0] - 2020-11-27
 
 ### Added